// src/app/modules/lshp-admin/models/institution-view.ts
import { InstitutionLicenseView } from './institution-license-view';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InstitutionView
 */
export interface InstitutionView extends InstitutionLicenseView {
     /**
     * Gets or sets Code
     */
    code: string;
    /**
     * Gets or sets Name
     */
    name: string;
    /**
     * Gets or sets Line1
     */
    line1: string;
    /**
     * Gets or sets Line2
     */
    line2?: string;
    /**
     * Gets or sets Line3
     */
    line3?: string;
    /**
     * Gets or sets City
     */
    city?: string;
    /**
     * Gets or sets StateId
     */
    stateId: number;
    /**
     * Gets or sets Pincode
     */
    pincode: string;
    /**
     * Gets or sets CountryId
     */
    countryId: number;
    /**
     * Gets or sets Website
     */
    website?: string;
    /**
     * Gets or sets Email1
     */
    email1: string;
    /**
     * Gets or sets Email2
     */
    email2?: string;
    /**
     * Gets or sets MobileTel
     */
    mobileTel: string;
    /**
     * Gets or sets LandTel
     */
    landTel?: string;
    /**
     * Gets or sets OtherTel
     */
    otherTel?: string;
    /**
     * Gets or sets UIThemeId
     */
    uiThemeId: number;
    
    institutionSymbolFileNameWithExtension: string;
    institutionSymbolFileValue: string;
    institutionNameFileNameWithExtension: string;
    institutionNameFileValue: string;
    /**
     * Gets or sets BillingCycleId
     */
    billingCycleId: number;
    /**
     * Gets or sets NumberOfSchools
     */
    numberOfSchools: number;
    /**
     * Gets or sets BillingStartDate, DateFormat MM-dd-YYYY
     */
    billingStartDate: Date;
    /**
     * Gets or sets BillingAmount
     */
    billingAmount: number;
}