// src/app/modules/lshp-admin/models/file-base64-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface FileBase64Model { 
    fileNameWithExtension?: string;
    fileType?: string;
    value?: string;
    name?: string;
}