// src/app/modules/lshp-admin/models/institution-update-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InstitutionUpdateView
 */
export interface InstitutionUpdateView { 
    /**
     * Gets or Sets code
     */
    code: string;
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or sets Name
     */
    name: string;
    /**
     * Gets or sets Line1
     */
    line1: string;
    /**
     * Gets or sets Line2
     */
    line2?: string;
    /**
     * Gets or sets Line3
     */
    line3?: string;
    /**
     * Gets or sets City
     */
    city?: string;
    /**
     * Gets or sets StateId
     */
    stateId: number;
    /**
     * Gets or sets
     */
    pincode: string;
    /**
     * Gets or sets
     */
    countryId: number;
    /**
     * Gets or sets
     */
    website?: string;
    /**
     * Gets or sets
     */
    email1: string;
    /**
     * Gets or sets
     */
    email2?: string;
    /**
     * Gets or sets
     */
    mobileTel: string;
    /**
     * Gets or sets
     */
    landTel?: string;
    /**
     * Gets or sets
     */
    otherTel?: string;
    /**
     * Gets or sets
     */
    uiThemeId: number;
    institutionSymbolFileNameWithExtension: string;
    institutionSymbolFileValue: string;
    institutionNameFileNameWithExtension: string;
    institutionNameFileValue: string;
}